third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/svg/text/scripted/getrotationofchar-expected.txt

This is a testharness.js-based test.
PASS 'text-orientation: mixed' - horizontal-tb
PASS 'text-orientation: upright' - horizontal-tb
PASS 'text-orientation: sideways' - horizontal-tb
FAIL Rotation attribute - horizontal-tb assert_equals: c expected 0 but got 360
PASS <textPath> - horizontal-tb
FAIL 'text-orientation: mixed' - vertical-rl assert_equals: 月 expected 0 but got 90
FAIL 'text-orientation: upright' - vertical-rl assert_equals: T expected 0 but got 90
PASS 'text-orientation: sideways' - vertical-rl
FAIL Rotation attribute - vertical-rl assert_equals: 木 expected 180.5 but got 270.5
FAIL <textPath> - vertical-rl assert_equals: 金 expected 315 but got 45
FAIL 'text-orientation: mixed' - vertical-lr assert_equals: 月 expected 0 but got 90
FAIL 'text-orientation: upright' - vertical-lr assert_equals: T expected 0 but got 90
PASS 'text-orientation: sideways' - vertical-lr
FAIL Rotation attribute - vertical-lr assert_equals: 木 expected 180.5 but got 270.5
FAIL <textPath> - vertical-lr assert_equals: 金 expected 315 but got 45
PASS 'text-orientation: mixed' - sideways-rl
PASS 'text-orientation: upright' - sideways-rl
PASS 'text-orientation: sideways' - sideways-rl
PASS Rotation attribute - sideways-rl
PASS <textPath> - sideways-rl
FAIL 'text-orientation: mixed' - sideways-lr assert_equals: M expected -90 but got 90
FAIL 'text-orientation: upright' - sideways-lr assert_equals: T expected -90 but got 90
FAIL 'text-orientation: sideways' - sideways-lr assert_equals: W expected -90 but got 90
FAIL Rotation attribute - sideways-lr assert_equals: t expected 270 but got 90
FAIL <textPath> - sideways-lr assert_equals: F expected 225 but got 45
Harness: the test ran to completion.
